+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - Replaced the arg_parser code with gen_class.ArgParser code.
 
 ### Changed
+- get_log: Refactored code.
 - Multiple functions: Replaced the arg_parser code with gen_class.ArgParser code.
 - main: Removed gen_libs.get_inst call.